Move `log_manager` under `crates/extensions/c8y_log_manager`
