+++ crates/extensions/c8y_config_manager/src/lib.rs
